requirements-api.txt
fastapi==0.110.1
uvicorn[standard]==0.29.0
python-multipart==0.0.9
python-dotenv==1.0.1
pydantic==2.7.0
pillow>=9.3.0
